Add a new zettelstack
The workflow it tries to enable is to insert links to thinks that needs
get explained so that I can fill the holes in a text later.

E.g.

Type in:
Foo is a variable name. The next should be Bar.

where Foo and Bar are links to new zettels. The next time
~zettelkasten-pop-new-file-stack~ is called it would visit Bar.org and
then Foo.org.